backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2 
pydantic[email]==2.10.4
resend==2.23.0
qrcode[pil]==7.4.2
# Security & Utilities
python-dotenv==1.0.1 
passlib[bcrypt]==1.7.4  
python-multipart==0.0.6  